Exclude databind from default test phase
